Adicionando Funcionalidade e crud completo de acomodações, e alterando algumas coisas de comodidade, para completar o acomodação

+++ src/main/java/com/unesc/earthBnb/service/AvaliacaoService.java
@@ -51,6 +51,9 @@ public class AvaliacaoService {
 
     @Transactional
     public void deleteAvaliacoes(Long id) {
+        if (!avaliacaoRepository.existsById(id)) {
+            throw new AvaliacaoNaoEncontradoException(id);
+        }
         avaliacaoRepository.deleteById(id);
     }
